Store irretrievable files in memory (#3746)
Pipes cannot be read multiple times, but RT relies on reading a file
at least twice to get the EXIF data. I made all file reads go through
`myfile.h` and added an `IrretrievableFileStore` singleton to recover
files on which `ftell()` failed from memory. To do so, I made the
`myfile.h` functions return and deal with `std::shared_ptr<IMFILE>`.
That would also open a possibility to add a cache for e.g. small files,
although then we would have to return new `IMFILE`s with a shared
`data` member, because `eof` and `pos` could be different.

This does not work for `/dev/stdin` right now, because we don't get a
thumbnail for it from `CacheManager::getEntry()`. But we need some
special handling for `STDIN` anyway.

For now, this is just a proof of concept.